start game button should close lobby options popup
