+++ src/WebFrontEnd/ClientApp/src/app/modules/lobby/components/gamemode-dialog/gamemode-dialog.component.ts
@@ -1,16 +1,16 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import {Subscription} from 'rxjs'
+import { API } from '@core/http/api';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import Lobby from '@core/models/lobby';
 import GameModes from '@core/models/gamemodes';
-import {Observable} from 'rxjs'
 import { ErrorService } from '@modules/lobby/services/error.service';
 
 interface GameModeDialogData {
   gameModes: GameModes.GameModeMetadata
   lobby: Lobby.LobbyMetadata
   error: string
-  onStart: () => Observable<any>
+  onGetLobby: () => void
 }
 
 @Component({
@@ -22,16 +22,17 @@ export class GamemodeDialogComponent implements OnInit {
   gameModes: GameModes.GameModeMetadata
   lobby: Lobby.LobbyMetadata
   error: string
+  onGetLobby: () => void
   errorSubscription: Subscription
-  onStart: () => Observable<any>
 
   constructor(
     private dialogRef: MatDialogRef<GamemodeDialogComponent>,
     public errorService: ErrorService,
-    @Inject(MAT_DIALOG_DATA) public data: GameModeDialogData) {
+    @Inject(MAT_DIALOG_DATA) public data: GameModeDialogData,
+    @Inject(API) private api: API) {
     this.gameModes = data.gameModes;
     this.lobby = data.lobby;
-    this.onStart = data.onStart;
+    this.onGetLobby = data.onGetLobby;
   }
 
   ngOnInit() {
@@ -44,12 +45,33 @@ export class GamemodeDialogComponent implements OnInit {
     this.errorSubscription.unsubscribe();
   }
 
-  onStartLobby = () => {
-    let request = this.onStart();
-    request.subscribe({
-      next: () => {
-        this.closeDialog();
-      }
+  onStartLobby() : void {
+    var body = new Lobby.ConfigureLobbyRequest();
+    body.gameMode = this.lobby.selectedGameMode;
+    body.options = JSON.parse(JSON.stringify(this.gameModes[this.lobby.selectedGameMode].options, ['value']));
+    var bodyString = JSON.stringify(body);
+
+    let configureRequest = this.api.request({ type: "Lobby", path: "Configure", body: bodyString })
+    let lobbyRequest = this.api.request({ type: "Lobby", path: "Start" })
+
+    configureRequest.subscribe({
+        next: () => {
+            lobbyRequest.subscribe({
+                next: () => { 
+                    this.onGetLobby()
+                    this.closeDialog();
+                },
+                error: (error) => { 
+                    this.error = error.error; 
+                    this.errorService.announceError(error.error); 
+                    this.onGetLobby() 
+                }
+            })
+        },
+        error: (error) => { 
+            this.error = error.error; 
+            this.onGetLobby(); 
+        }
     })
   }
